backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.17.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
filelock==3.15.1
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
llama-index==0.10.45
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.44
llama-index-embeddings-azure-openai==0.1.10
llama-index-embeddings-huggingface==0.2.1
llama-index-embeddings-openai==0.1.10
llama-index-experimental==0.1.3
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.1.8
llama-index-llms-openai==0.1.22
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-azurecosmosmongo==0.1.3
llama-parse==0.4.4
llamaindex-py-client==0.1.19
MarkupSafe==2.1.5
marshmallow==3.21.3
minijinja==2.0.1
mkl==2021.4.0
mpmath==1.3.0
msal==1.29.0
msal-extensions==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==2.0.0
openai==1.34.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pip==24.0
portalocker==2.10.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pymongo==4.7.3
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.7.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
striprtf==0.0.26
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
